+++ Examples/Example-UIKit/Example-UIKit/Application/SceneDelegate.swift
@@ -1,3 +1,4 @@
+// BTT_INJECTED: SceneDelegate.swift
 //
 //  SceneDelegate.swift
 //  Example-UIKit
@@ -6,6 +7,7 @@
 //
 
 import UIKit
+import SwiftUI
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
@@ -23,6 +25,12 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window.rootViewController = navigationController
         window.makeKeyAndVisible()
         self.window = window
+        
+       /* let window = UIWindow(windowScene: windowScene)
+        window.rootViewController = UIHostingController(rootView: DemoView())
+        window.makeKeyAndVisible()
+        self.window = window*/
+
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {