+++ Examples/Example-UIKit/Example-UIKit/ShipView.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct ShipView: View {
+    var body: some View {
+        VStack {
+            Text("Ship")
+        }
+    }
+}
+
+