+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Blue Triangle 3.15.12, Latest
+### New Features and Bug Fixes
+- Added .bttTrack modifier for auto instrumentor tracking
+- Added NavigationStackTracker for NavigationStack path tracking
+- Fixed multiple group creation for the same screen during scrolling
+- Fixed impact on screen name resolution caused by repeated child views
+
 # Blue Triangle 3.15.12, May 22 2026
 ### New Features and Bug Fixes
 - App installs tracking: BlueTriangle can now track new app installs.